Adapt ML Benchmark Template to fit with current API
