+++ hpobench/dependencies/ml/ml_benchmark_template.py
@@ -1,0 +1,277 @@
+import os
+import time
+import openml
+import numpy as np
+import pandas as pd
+import ConfigSpace as CS
+from typing import Union, Dict
+from pathlib import Path
+
+from sklearn.metrics import make_scorer, accuracy_score, balanced_accuracy_score, \
+    precision_score, f1_score
+
+from hpobench.abstract_benchmark import AbstractBenchmark
+from hpobench.dependencies.ml.data_manager import OpenMLDataManager
+
+metrics = dict(
+    acc=accuracy_score,
+    bal_acc=balanced_accuracy_score,
+    f1=f1_score,
+    precision=precision_score,
+)
+
+metrics_kwargs = dict(
+    acc=dict(),
+    bal_acc=dict(),
+    f1=dict(average="macro", zero_division=0),
+    precision=dict(average="macro", zero_division=0),
+)
+
+
+class MLBenchmark(AbstractBenchmark):
+    _issue_tasks = [3917, 3945]
+
+    def __init__(
+            self,
+            task_id: Union[int, None] = None,
+            rng: Union[np.random.RandomState, int, None] = None,
+            valid_size: float = 0.33,
+            data_path: Union[str, Path, None] = None,
+            global_seed: int = 1
+    ):
+        super(MLBenchmark, self).__init__(rng=rng)
+
+        if isinstance(rng, int):
+            self.seed = rng
+        else:
+            self.seed = self.rng.randint(1, 10**6)
+
+        self.global_seed = global_seed  # used for fixed training-validation splits
+
+        self.task_id = task_id
+        self.valid_size = valid_size
+        self.scorers = dict()
+        for k, v in metrics.items():
+            self.scorers[k] = make_scorer(v, **metrics_kwargs[k])
+        self.data_path = data_path
+
+        dm = OpenMLDataManager(task_id, valid_size, data_path, global_seed)
+        dm.load()
+
+        # Data variables
+        self.train_X = dm.train_X
+        self.valid_X = dm.valid_X
+        self.test_X = dm.test_X
+        self.train_y = dm.train_y
+        self.valid_y = dm.valid_y
+        self.test_y = dm.test_y
+        self.train_idx = dm.train_idx
+        self.test_idx = dm.test_idx
+        self.task = dm.task
+        self.dataset = dm.dataset
+        self.preprocessor = dm.preprocessor
+        self.lower_bound_train_size = dm.lower_bound_train_size
+        self.n_classes = dm.n_classes
+
+        # Observation and fidelity spaces
+        self.fidelity_space = self.get_fidelity_space(self.seed)
+        self.configuration_space = self.get_configuration_space(self.seed)
+
+    @staticmethod
+    def get_configuration_space(seed=None):
+        """Parameter space to be optimized --- contains the hyperparameters
+        """
+        raise NotImplementedError()
+
+    @staticmethod
+    def get_fidelity_space(seed=None):
+        """Fidelity space available --- specifies the fidelity dimensions
+
+        If fidelity_choice is 0
+            Fidelity space is the maximal fidelity, akin to a black-box function
+        If fidelity_choice is 1
+            Fidelity space is a single fidelity, in this case the number of trees (n_estimators)
+        If fidelity_choice is 2
+            Fidelity space is a single fidelity, in this case the fraction of dataset (subsample)
+        If fidelity_choice is >2
+            Fidelity space is multi-multi fidelity, all possible fidelities
+        """
+        raise NotImplementedError()
+
+    def get_meta_information(self):
+        """ Returns the meta information for the benchmark """
+        return {
+            'name': 'Support Vector Machine',
+            'shape of train data': self.train_X.shape,
+            'shape of test data': self.test_X.shape,
+            'shape of valid data': self.valid_X.shape,
+            'initial random seed': self.seed,
+            'task_id': self.task_id
+        }
+
+    def init_model(self, config, fidelity=None, rng=None):
+        """ Function that returns the model initialized based on the configuration and fidelity
+        """
+        raise NotImplementedError()
+
+    def get_config(self, size=None):
+        """Samples configuration(s) from the (hyper) parameter space
+        """
+        if size is None:  # return only one config
+            return self.configuration_space.sample_configuration()
+        return [self.configuration_space.sample_configuration() for i in range(size)]
+
+    def get_fidelity(self, size=None):
+        """Samples candidate fidelities from the fidelity space
+        """
+        if size is None:  # return only one config
+            return self.fidelity_space.sample_configuration()
+        return [self.fidelity_space.sample_configuration() for i in range(size)]
+
+    def shuffle_data_idx(self, train_idx=None, rng=None):
+        rng = self.rng if rng is None else rng
+        train_idx = self.train_idx if train_idx is None else train_idx
+        rng.shuffle(train_idx)
+        return train_idx
+
+    def _train_objective(self, config, fidelity, shuffle, rng, evaluation="valid"):
+        # initializing model
+        model = self.init_model(config, fidelity, rng)
+
+        # preparing data
+        if eval == "valid":
+            train_X = self.train_X
+            train_y = self.train_y
+            train_idx = self.train_idx
+        else:
+            train_X = np.vstack((self.train_X, self.valid_X))
+            train_y = pd.concat((self.train_y, self.valid_y))
+            train_idx = np.arange(len(train_X))
+
+        # shuffling data
+        if shuffle:
+            train_idx = self.shuffle_data_idx(train_idx, rng)
+            train_X = train_X.iloc[train_idx]
+            train_y = train_y.iloc[train_idx]
+
+        # subsample here:
+        # application of the other fidelity to the dataset that the model interfaces
+        if self.lower_bound_train_size is None:
+            self.lower_bound_train_size = (10 * self.n_classes) / self.train_X.shape[0]
+            self.lower_bound_train_size = np.max((1 / 512, self.lower_bound_train_size))
+        subsample = np.max((fidelity['subsample'], self.lower_bound_train_size))
+        train_idx = self.rng.choice(
+            np.arange(len(train_X)), size=int(
+                subsample * len(train_X)
+            )
+        )
+        # fitting the model with subsampled data
+        start = time.time()
+        model.fit(train_X[train_idx], train_y.iloc[train_idx])
+        model_fit_time = time.time() - start
+        # computing statistics on training data
+        scores = dict()
+        score_cost = dict()
+        for k, v in self.scorers.items():
+            scores[k] = 0.0
+            score_cost[k] = 0.0
+            if evaluation == "test":
+                _start = time.time()
+                scores[k] = v(model, train_X, train_y)
+                score_cost[k] = time.time() - _start
+        train_loss = 1 - scores["acc"]
+        return model, model_fit_time, train_loss, scores, score_cost
+
+    # pylint: disable=arguments-differ
+    @AbstractBenchmark.check_parameters
+    def objective_function(self,
+                           configuration: Union[CS.Configuration, Dict],
+                           fidelity: Union[CS.Configuration, Dict, None] = None,
+                           shuffle: bool = False,
+                           rng: Union[np.random.RandomState, int, None] = None,
+                           **kwargs) -> Dict:
+        """Function that evaluates a 'config' on a 'fidelity' on the validation set
+        """
+        model, model_fit_time, train_loss, train_scores, train_score_cost = self._train_objective(
+            configuration, fidelity, shuffle, rng, evaluation="val"
+        )
+        val_scores = dict()
+        val_score_cost = dict()
+        for k, v in self.scorers.items():
+            _start = time.time()
+            val_scores[k] = v(model, self.valid_X, self.valid_y)
+            val_score_cost[k] = time.time() - _start
+        val_loss = 1 - val_scores["acc"]
+
+        test_scores = dict()
+        test_score_cost = dict()
+        for k, v in self.scorers.items():
+            _start = time.time()
+            test_scores[k] = v(model, self.test_X, self.test_y)
+            test_score_cost[k] = time.time() - _start
+        test_loss = 1 - test_scores["acc"]
+
+        info = {
+            'train_loss': train_loss,
+            'val_loss': val_loss,
+            'test_loss': test_loss,
+            'model_cost': model_fit_time,
+            'train_scores': train_scores,
+            'train_costs': train_score_cost,
+            'val_scores': val_scores,
+            'val_costs': val_score_cost,
+            'test_scores': test_scores,
+            'test_costs': test_score_cost,
+            # storing as dictionary and not ConfigSpace saves tremendous memory
+            'fidelity': fidelity.get_dictionary(),
+            'config': configuration.get_dictionary()
+        }
+
+        return {
+            'function_value': info['val_loss'],
+            'cost': model_fit_time + info['val_costs']['acc'],
+            'info': info
+        }
+
+    # pylint: disable=arguments-differ
+    @AbstractBenchmark.check_parameters
+    def objective_function_test(self,
+                                configuration: Union[CS.Configuration, Dict],
+                                fidelity: Union[CS.Configuration, Dict, None] = None,
+                                shuffle: bool = False,
+                                rng: Union[np.random.RandomState, int, None] = None,
+                                **kwargs) -> Dict:
+        """Function that evaluates a 'config' on a 'fidelity' on the test set
+        """
+        model, model_fit_time, train_loss, train_scores, train_score_cost = self._train_objective(
+            configuration, fidelity, shuffle, rng, evaluation="test"
+        )
+        test_scores = dict()
+        test_score_cost = dict()
+        for k, v in self.scorers.items():
+            _start = time.time()
+            test_scores[k] = v(model, self.test_X, self.test_y)
+            test_score_cost[k] = time.time() - _start
+        test_loss = 1 - test_scores["acc"]
+
+        info = {
+            'train_loss': train_loss,
+            'val_loss': None,
+            'test_loss': test_loss,
+            'model_cost': model_fit_time,
+            'train_scores': train_scores,
+            'train_costs': train_score_cost,
+            'val_scores': dict(),
+            'val_costs': dict(),
+            'test_scores': test_scores,
+            'test_costs': test_score_cost,
+            # storing as dictionary and not ConfigSpace saves tremendous memory
+            'fidelity': fidelity.get_dictionary(),
+            'config': configuration.get_dictionary()
+        }
+
+        return {
+            'function_value': info['test_loss'],
+            'cost': model_fit_time + info['test_costs']['acc'],
+            'info': info
+        }
